+++ org.mwc.debrief.dis/src/org/mwc/debrief/dis/listeners/IDISFixListener.java
@@ -2,6 +2,6 @@ package org.mwc.debrief.dis.listeners;
 
 public interface IDISFixListener
 {
-  void add(long time, long exerciseId, long id, double dLat, double dLong,
+  void add(long time, short exerciseId, long id, double dLat, double dLong,
       double depth, double courseDegs, double speedMS);
 }
